+++ src/x86_64_sub.c
@@ -26,7 +26,7 @@ x86_64_sub_r64_imm8(X86_64_REG reg,
                     int8_t value)
 {
 	return x86_64_rex(1, 0, 0, X86_64_REG64(reg)) +
-	       x86_64_write(0x83) +
+	       x86_64_write((int8_t) 0x83) +
 	       x86_64_write(0xc0 + 0x28 + (X86_64_REG64(reg) ? reg - 8 : reg)) +
 	       x86_64_write(value);
 }
@@ -38,7 +38,7 @@ x86_64_sub_r32_imm32(X86_64_REG reg,
 	return x86_64_rex(1, 0, 0, X86_64_REG64(reg)) +
 	       (reg == X86_64_RAX ? 
 	           x86_64_write(0x05 + 0x28) :
-	           x86_64_write(0x81) +
+	           x86_64_write((int8_t) 0x81) +
 	           x86_64_write(0xc0 + 0x28 + (X86_64_REG64(reg) ? reg - 8 : reg))) +
 	       x86_64_imm32(value);
 }
@@ -57,7 +57,7 @@ x86_64_sub_rm64_imm8(X86_64_REG reg,
                      int8_t value)
 {
 	return (X86_64_REG64(reg) ? x86_64_rex(0, 0, 0, 1) : 0) +
-	       x86_64_write(0x80) +
+	       x86_64_write((int8_t) 0x80) +
 	       x86_64_write(0x28 + (X86_64_REG64(reg) ? reg - 8 : reg)) +
 	       x86_64_write(value);
 }
